feat: Write output.js if existing
